Make a preference to disable loading Sparks.
Could be a memory hog, even if people are not reading them. I
personally only keep them only for Hot Items, and still they make
up the major part of my downloads.

This change won't disable the download, but the saving of Sparks. I'm
not sure if it's possible with the Fever API to disable loading of 
Sparks at the lowest level...